Fix for size_t coders in 32-bit build
We need to encode/decode the size of a Vector, which is of unsigned long
type, but we don't get it from any of the definitions in 32-bit builds.

+++ Source/WebCore/platform/WebCorePersistentCoders.cpp
@@ -425,7 +425,7 @@ void Coder<WebCore::CertificateInfo>::encode(Encoder& encoder, const WebCore::Ce
     auto& certificateChain = certificateInfo.certificateChain();
 
     encoder << certificateInfo.verificationError();
-    encoder << certificateChain.size();
+    encoder << (uint64_t)certificateChain.size();
     for (auto& certificate : certificateChain)
         encoder << certificate;
 }
@@ -437,7 +437,7 @@ std::optional<WebCore::CertificateInfo> Coder<WebCore::CertificateInfo>::decode(
     if (!verificationError)
         return std::nullopt;
 
-    std::optional<size_t> numOfCerts;
+    std::optional<uint64_t> numOfCerts;
     decoder >> numOfCerts;
     if (!numOfCerts)
         return std::nullopt;
